Convert nfa to dfa
Turn a nondeterministic finite automaton into a deterministic one by
subset construction. The main algorithm seems to work. We need a way
to unit test it.

There are several things to work on next. We have to remember which
state is the implicitly constructed explicit error state. And for each
accepting state we have to remember which token it accepts.