Remember useful information in sigil::Grammar
We want to do some things with the resulting `sigil::Grammar`. For
example using the created dfa to generate a scanner. This of course
requires us to remember some information.

In this patch we remember the dfa and the names of the tokens. This
allows us to black-box unit-test the algorithm creating automatons.