Implement a scanner driver
The scanner driver is implemented as a abstract class with the same
name (`ScannerDriver`). It provides a generic interface and
implementation for different dfa representations.

Currently subclass `sigil::DfaScannerDriver` uses
`sigil::dfa::Automaton` to implement the interface. This is not the
most efficient implementation since `sigil::DfaScannerDriver` has to
search to find the next applicable state. A table driven
implementation would elimiate the need for searching.