Let the user pass in a token type
Previously the accepted token type was coupled to the order in which
the tokens were registered. This made no real sense.

Now the ordering is used for resolving ambiguities (e.g. if you have
keyword 'if' and also identifiers '[a-zA-Z]+'), but the returned value
can now be controlled by the user.