Fix can_be_atom
Predicate `can_be_atom` was supposed to return true, iff a character
can stand for its own. This would be not the case for meta-characters,
such as '|' (alternative), '(' and ')' (parentheses for groups), '['
and ']' (brackets for character classes), '{' and '}' (for repetition
- currently unsupported by sigil) and '.' (wildcard character).

So we turn the predicate around by introducing a predicate for meta
characters (`is_meta_character`) and use it's negation to implement
`can_be_atom`.

I think the implemented behavior is still incorrect, since '.' can
stand on its own. Depending on the context it refers to a different
set of characters, though: If it's part of a top-level delaration, the
wildcard is meant, it's part of a character class, the literal
character '.' is meant.

Still, this is an improvement, since it allows us to specify the regex
"\\[^\n]*\n" which would previously fail to parse at position 4 (the
"\n" inside the character classs).