Emit correct instance of `sigil::StaticTable`
We previously used a huge integer array-literal to emit transitions
and accepting states for the generated term when printing
`sigil::StaticTable`.

However there was a problem with that. The generated table was
sometimes not the same as the source table.

We now use a new datastructure `sigil::Array`, which may be
initialized by a string-literal. This fixes encoding the table and
also speeds up compilation of generated tables as a nice side-effect.